pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marcus-core"
version = "1.0.0"
description = "OSINT-enabled facial analysis system with YOLO detection, deep embeddings, and vector search"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MARCUS Project"}
]
keywords = ["face-recognition", "osint", "biometrics", "yolo", "arcface", "faiss"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "hnswlib>=0.7.0",
    "insightface>=0.7.3",
    "onnxruntime>=1.16.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "scikit-learn>=1.3.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.28.0",
    "streamlit-webrtc>=0.45.0",
    "av>=10.0.0",
]
gpu = [
    "onnxruntime-gpu>=1.16.0",
    "faiss-gpu>=1.7.4",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]
all = [
    "marcus-core[ui,dev]",
]

[project.scripts]
marcus = "marcus_core.cli:main"

[project.urls]
Homepage = "https://github.com/marcus-project/marcus-core"
Documentation = "https://marcus-project.github.io/marcus-core"
Repository = "https://github.com/marcus-project/marcus-core"

[tool.setuptools.packages.find]
where = ["."]
include = ["marcus_core*", "marcus_ui*"]

[tool.setuptools.package-data]
marcus_core = ["*.yaml", "*.json"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | models
    | data
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=marcus_core --cov-report=term-missing"
asyncio_mode = "auto"
